Refactored MarketplaceFeed/Utils
