+++ components/Marketplace/MarketplaceFeed.tsx
@@ -1,10 +1,10 @@
 import { useState } from 'react';
 import { Group, Loader, SimpleGrid, Text } from '@mantine/core';
-import { ItemForSale } from '@/src/API';
-import { MarketplaceCard } from './MarketplaceCard';
-import { ViewListingModal } from './ViewListingModal';
+import { ItemForSale } from '@/types/types';
+import { MarketplaceCard } from '@/components/Marketplace/MarketplaceCard';
+import { ViewListingModal } from '@/components/Marketplace/ViewListingModal';
 import { useFetchListings } from '@/src/hooks/marketplaceCustomHooks';
-import { sortByCreatedAt, sortByPriceHighLow, sortByPriceLowHigh } from './marketplaceSort';
+import { filterAndSortListings } from '@/components/utils/marketplaceUtils';
 
 export function MarketplaceFeed({
   refresh,
@@ -18,36 +18,12 @@ export function MarketplaceFeed({
   const { listings, loading } = useFetchListings(refresh);
   const [viewListingModalOpened, setViewListingModalOpened] = useState(false);
   const [selectedListing, setSelectedListing] = useState(null);
+  const filteredAndSortedListings = filterAndSortListings(listings, searchQuery, sortQuery);
 
   const handleViewListing = (item: any) => {
     setSelectedListing(item);
     setViewListingModalOpened(true);
   };
-
-  const filteredListings = listings.filter(
-    (item: ItemForSale) =>
-      item.title.toLowerCase().includes(searchQuery.toLowerCase()) ||
-      item.description.toLowerCase().includes(searchQuery.toLowerCase()) ||
-      `${item.seller.firstName} ${item.seller.lastName}`
-        .toLowerCase()
-        .includes(searchQuery.toLowerCase())
-  );
-
-  let sortedListings: ItemForSale[] = [];
-  switch (sortQuery) {
-    case 'Price: Low to High':
-      sortedListings = filteredListings.sort(sortByPriceLowHigh);
-      break;
-    case 'Price: High to Low':
-      sortedListings = filteredListings.sort(sortByPriceHighLow);
-      break;
-    case 'Newly Listed':
-      sortedListings = filteredListings.sort(sortByCreatedAt);
-      break;
-    default:
-      sortedListings = filteredListings;
-      break;
-  }
 
   return (
     <>
@@ -61,7 +37,7 @@ export function MarketplaceFeed({
             No item is up for grabs yet, try listing yours!
           </Text>
         </Group>
-      ) : filteredListings.length === 0 ? (
+      ) : filteredAndSortedListings.length === 0 ? (
         <Group justify="center" mt="200">
           <Text size="xl" c="dimmed">
             There is no item that matches your search query
@@ -74,7 +50,7 @@ export function MarketplaceFeed({
           verticalSpacing={{ base: 'md', sm: 'lg' }}
           data-testid="marketplace-feed"
         >
-          {sortedListings.map((item: ItemForSale) => (
+          {filteredAndSortedListings.map((item: ItemForSale) => (
             <MarketplaceCard key={item.id} item={item} onView={() => handleViewListing(item)} />
           ))}
         </SimpleGrid>
